include audioconvert, audioconfig in GStreamer pipeline
> Brian, I will release libcanberra 0.8 shortly. If you want to make
> sure that the Gst driver works fine on solaris before I do the
> release, please test quickly!

Sorry I was not able ot test before the release.  I just tested, and
I found it doesn't work for two reasons.  One reason I fixed in the
attached patch.  On Solaris, it is necessary to include "audioconvert !
audioconfig" in the GStreamer pipeline.  This is needed to ensure that
GStreaner "just works" with sink plugins that may not implement the
full range of possible output that can come out of decodebin.

The attached patch fixes this.  Totem, rhythmbox, and other GStreamer
based programs use similar logic.

Before the above fix, GStreamer would just hang, which would also
happen with the gst-launch command if you ran it with a similar
pipeline without "audioconvert".

Brian

Signed-off-by: Marc-André Lureau <marcandre.lureau@gmail.com>
Signed-off-by: Lennart Poettering <lennart@poettering.net>