Fixes for libcanberra GStreamer backend
I wrote the attached patch to fix some bugs in the GStreamer backend of
libcanberra.

When you run gnome-sound-properties, and then enable "Play alerts and
sound effects" on the "Sounds" tab, then click on one of the little
arrow buttons to preview the sound, it plays okay.  However, the second
time you try clicking on any little arrow preview buttons it crashes.
It does this even if you wait for the first sound to stop playing before
playing the second sound.

Signed-off-by: Lennart Poettering <lennart@poettering.net>