pulse: fix a minor race with sound cancellation
If we call pa_stream_drain() and immediately afterwards destroy the
stream it might happen that the pa_stream_drain() callback is called
after the destruction finished pointing to an invalid out structure.

To fix this we need to terminate the _drain() callback to make sure the
reference to the not existing structure is dropped.